# src/ui/openapi.yaml
openapi: 3.0.3
info:
  version: 1.3.42
  description: |
    REST API surface served by warest-whatsapp-rest-api-multisessions.
  contact:
    name: WARest
servers:
  - url: http://localhost:4001
    description: Local development
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/messages/send/media:
    post:
      operationId: postMessagesSendMedia
      tags:
        - Messages - Sending
      summary: Send an image or video
      description: |
        Send image/video by HTTPS URL, base64 data URI, or direct file upload (multipart/form-data).
        Attach optional caption, compression, and view once flags.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  required:
                    - media
                  properties:
                    media:
                      type: string
                      example: "https://example.com/photo.jpg"
                      description: HTTPS URL, local path, or base64 data URI.
                    caption:
                      type: string
                    compress:
                      type: boolean
                      default: true
                    viewOnce:
                      type: boolean
                      default: false
                - $ref: "#/components/schemas/SendOptions"
            example:
              sessionId: main
              to: "62812xxxxxxx"
              media: "https://example.com/photo.jpg"
              caption: "Hi!"
              compress: true
              viewOnce: false
              presence: true
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  properties:
                    media:
                      type: string
                      format: binary
                      description: Binary image/video file.
                    caption:
                      type: string
                      description: (optional) Caption for the media.
                    compress:
                      type: boolean
                      default: true
                      description: (optional) Whether to compress the media.
                    viewOnce:
                      type: boolean
                      default: false
                      description: (optional) Send as view once media.
                  required:
                    - media
                - $ref: "#/components/schemas/SendOptions"
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: >
              Raw image/video bytes. Supply `sessionId`/`phone`, `to`, and optional send options via query string.
      responses:
        <<: *commonResponses
        200:
          description: Media message accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSendResponse"
              examples:
                mediaSuccess:
                  summary: Image message sent
                  value:
                    status: true
                    code: 1000
                    message: "Media message sent"
                    results:
                      - phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB07890..."
                        status: "image success"
                        timestamp: 1763487010
                        advancedInfo:
                          isForwarded: false
                          quotedMessageId: null
                          compression:
                            enable: true
                          mediaKind: image
  /api/v1/messages/send/audio:
    post:
      operationId: postMessagesSendAudio
      tags:
        - Messages - Sending
      summary: Send an audio message
      description: Send audio or push-to-talk voice note.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  required:
                    - audio
                  properties:
                    audio:
                      type: string
                      description: URL or base64 payload.
                    isVN:
                      type: boolean
                      description: Send as voice note (PTT) when true.
                - $ref: "#/components/schemas/SendOptions"
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  properties:
                    audio:
                      type: string
                      format: binary
                      description: Binary audio upload.
                    isVN:
                      type: boolean
                  required:
                    - audio
                - $ref: "#/components/schemas/SendOptions"
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: >
              Raw audio bytes. Provide `sessionId`/`phone`, `to`, and optional options via query string parameters.
            example:
              sessionId: main
              to: "62812xxxxxxx"
              audio: "data:audio/ogg;base64,..."
              isVN: true
              presence: true
      responses:
        <<: *commonResponses
        200:
          description: Audio dispatch summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSendResponse"
              examples:
                audioSuccess:
                  summary: Voice note sent
                  value:
                    status: true
                    code: 1000
                    message: "Audio message sent"
                    results:
                      - phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB0A111..."
                        status: "ptt success"
                        timestamp: 1763487020
                        advancedInfo:
                          isForwarded: false
                          quotedMessageId: string
                          isVN: true
                          mimeType: "audio/ogg; codecs=opus"
  /api/v1/messages/send/document:
    post:
      operationId: postMessagesSendDocument
      tags:
        - Messages - Sending
      summary: Send a document
      description: |
        Send document or any generic file as document message. Archives (ZIP/7z), PDFs, spreadsheets, and
        other binary formats are accepted, while executables and scripts are blocked automatically.
        MIME validation follows `WAREST_MIMETYPE_FILES_ALLOWLIST` with the same defaults as `/messages/send/files`:
          - image_as_document: image/jpeg, image/png, image/webp, image/gif, image/bmp, image/tiff, image/heif, image/heic
          - video_as_document: video/mp4, video/3gpp, video/3gp, video/mpeg, video/ogg, video/webm, video/x-msvideo, video/x-matroska, video/quicktime
          - audio_as_document: audio/mpeg, audio/mp3, audio/aac, audio/wav, audio/flac, audio/ogg, audio/opus, audio/amr, audio/x-ms-wma, audio/mp4, audio/m4a
          - office_documents: application/pdf, application/msword, application/vnd.ms-excel, application/vnd.ms-powerpoint, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.openxmlformats-officedocument.presentationml.presentation, application/vnd.ms-access, application/vnd.visio, application/vnd.ms-project
          - archives_and_packages: application/zip, application/x-zip-compressed, application/vnd.rar, application/x-rar-compressed, application/gzip, application/x-gzip, application/x-7z-compressed, application/x-tar, application/x-bzip2, application/x-xz, application/vnd.android.package-archive, application/x-msdownload, application/x-iso9660-image
          - text_and_data: text/plain, text/csv, text/xml, application/xml, application/json, application/javascript, text/html, text/markdown, application/sql, application/x-sql, text/yaml, application/x-yaml, application/x-sh, text/css
          - ebooks: application/epub+zip, application/x-mobipocket-ebook, application/x-fictionbook+xml
          - design_and_vector: application/postscript, image/svg+xml, application/illustrator, application/photoshop, application/vnd.adobe.photoshop, image/vnd.adobe.photoshop
          - fallback: application/octet-stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  required:
                    - document
                  properties:
                    document:
                      type: string
                      example: "https://example.com/contract.pdf"
                      description: URL, path, or base64 data URI.
                    filename:
                      type: string
                      example: "contract.pdf"
                      description: Filename of the document.
                    caption:
                      type: string
                      description: (optional) Document caption.
                - $ref: "#/components/schemas/SendOptions"
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  properties:
                    document:
                      type: string
                      format: binary
                      description: Binary document upload.
                    filename:
                      type: string
                    caption:
                      type: string
                  required:
                    - document
                - $ref: "#/components/schemas/SendOptions"
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: >
              Raw document bytes. Pass `sessionId`/`phone`, `to`, filename, and other options through the query string.
            example:
              sessionId: main
              to: "62812xxxxxxx"
              document: "https://example.com/contract.pdf"
              filename: "contract.pdf"
              caption: "Please review"
              presence: true
      responses:
        <<: *commonResponses
        200:
          description: Document message info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSendResponse"
              examples:
                documentSuccess:
                  summary: Document sent
                  value:
                    status: true
                    code: 1000
                    message: "Document message sent"
                    results:
                      - phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB0B222..."
                        status: "document success"
                        timestamp: 1763487030
                        advancedInfo:
                          isForwarded: false
                          quotedMessageId: string
                          fileName: "contract.pdf"
                          mimetype: "application/pdf"
  /api/v1/messages/send/sticker:
    post:
      operationId: postMessagesSendSticker
      tags:
        - Messages - Sending
      summary: Send a sticker
      description: Send sticker (converted to webp when necessary).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  required:
                    - sticker
                  properties:
                    sticker:
                      type: string
                      description: URL, base64 data URI, or storage path.
                    compress:
                      type: boolean
                      description: Enable resizing before converting to WebP.
                - $ref: "#/components/schemas/SendOptions"
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  properties:
                    sticker:
                      type: string
                      format: binary
                      description: Binary image/GIF upload.
                    compress:
                      type: boolean
                      description: Enable resizing before converting to WebP.
                  required:
                    - sticker
                - $ref: "#/components/schemas/SendOptions"
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: >
              Raw sticker bytes. Supply `sessionId`/`phone`, `to`, and additional options via query parameters.
            example:
              sessionId: main
              to: "62812xxxxxxx"
              sticker: "https://example.com/cat.png"
              presence: true
      responses:
        <<: *commonResponses
        200:
          description: Sticker send summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSendResponse"
              examples:
                stickerSuccess:
                  summary: Sticker sent
                  value:
                    status: true
                    code: 1000
                    message: "Sticker message sent"
                    results:
                      - phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB0C333..."
                        status: "sticker success"
                        timestamp: 1763487040
                        advancedInfo:
                          isForwarded: false
                          quotedMessageId: string
                          compression:
                            enable: true
...
  /api/v1/messages/send/gif:
    post:
      operationId: postMessagesSendGif
      tags:
        - Messages - Sending
      summary: Send a GIF
      description: Send GIF (converted to mp4 under the hood).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  required:
                    - gif
                  properties:
                    gif:
                      type: string
                      description: URL, base64 data URI, or storage path.
                    caption:
                      type: string
                      description: (optional) GIF caption.
                    compress:
                      type: boolean
                      description: Compress video/image payloads when possible.
                - $ref: "#/components/schemas/SendOptions"
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  properties:
                    gif:
                      type: string
                      format: binary
                      description: GIF/video upload.
                    caption:
                      type: string
                      description: (optional) GIF caption.
                    compress:
                      type: boolean
                      description: Compress video payloads when possible.
                  required:
                    - gif
                - $ref: "#/components/schemas/SendOptions"
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: >
              Raw GIF/video bytes. Provide `sessionId`/`phone`, `to`, and additional options via the query string.
            example:
              sessionId: main
              to: "62812xxxxxxx"
              gif: "https://example.com/loop.mp4"
              caption: string
              compress: true
              presence: true
      responses:
        <<: *commonResponses
        200:
          description: GIF send summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSendResponse"
              examples:
                gifSuccess:
                  summary: GIF message sent
                  value:
                    status: true
                    code: 1000
                    message: "GIF message sent"
                    results:
                      - phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB0F666..."
                        status: "gif success"
                        timestamp: 1763487070
                        advancedInfo:
                          isForwarded: false
                          quotedMessageId: string
                          caption: string
                          compression:
                            enable: true
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-WAREST-API-KEY
      description: "Send your X-WAREST-API-KEY header; use an admin-scoped key for admin-only routes."
  schemas:
    ApiEnvelope:
      type: object
      properties:
        status:
          type: boolean
        code:
          type: integer
        message:
          type: string
        results:
          type: array
          nullable: true
          items:
            type: object
      required: [status, code, message]
...
    SessionSelector:
      description: Provide the session handle by supplying `sessionId`. This identifier is returned by `/api/v1/session/create`.
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: string
          description: Unique session identifier returned by `/api/v1/session/create`.
          example: "main"
    DestinationSelector:
      allOf:
        - $ref: "#/components/schemas/SessionSelector"
        - type: object
          required:
            - to
          properties:
            to:
              oneOf:
                - type: string
                  description: Digit string, LID, or JID. Comma separated lists are allowed.
                  example: "6281234567890"
                - type: array
                  items:
                    type: string
    SendOptions:
      type: object
      properties:
        presence:
          type: boolean
          description: "(optional) When `true`, the server automatically marks the sender available/online and emits typing or recording cues (depending on the route) before sending the payload."
          example: true
        replyMessageId:
          type: string
          description: "(optional) Message ID to quote."
        isForwarded:
          type: boolean
          example: false
          description: "(optional) Mark the outgoing message as forwarded."
...
    MessageSendResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            results:
              type: array
              description: Array of per-recipient send results (empty when no recipient was processed).
              items:
                $ref: "#/components/schemas/SendResult"
      example:
        status: true
        code: 1000
        message: Text message sent
        results:
          - phone: "6281912311111"
            to: "628123456789"
            toJid: "628123456789@s.whatsapp.net"
            messageId: "3EB055069C144F12C0E64D"
            status: "text success"
            timestamp: 1763486858
            advancedInfo:
              isForwarded: false
              quotedMessageId: "ABCD1234EFGH5678"
...